+++ sila_cetoni/motioncontrol/axis/sila/axis_service/feature_implementations/axissystemcontrolservice_impl.py
@@ -1,17 +1,14 @@
 from __future__ import annotations
 
 import logging
-import time
-from concurrent.futures import Executor
-from threading import Event
-from typing import Any, Dict, List
+from typing import Dict, List
 
 from qmixsdk.qmixmotion import Axis, AxisSystem
-from sila2.framework import FullyQualifiedIdentifier
 from sila2.server import MetadataDict, SilaServer
 
 from sila_cetoni.application.server_configuration import ServerConfiguration
 from sila_cetoni.application.system import ApplicationSystem
+from sila_cetoni.utils import PropertyUpdater, not_equal
 
 from ..generated.axissystemcontrolservice import (
     AxisSystemControlServiceBase,
@@ -28,41 +25,29 @@ class AxisSystemControlServiceImpl(AxisSystemControlServiceBase):
     __axis_system: AxisSystem
     __axes: Dict[str, Axis]
     __config: ServerConfiguration
-    __stop_event: Event
 
-    def __init__(self, server: SilaServer, axis_system: AxisSystem, executor: Executor):
+    def __init__(self, server: SilaServer, axis_system: AxisSystem):
         super().__init__(server)
         self.__axis_system = axis_system
         self.__axes = {
             self.__axis_system.get_axis_device(i).get_device_name(): self.__axis_system.get_axis_device(i)
             for i in range(self.__axis_system.get_axes_count())
         }
-        self.__stop_event = Event()
 
-        def update_axis_system_state(stop_event: Event):
-            new_is_enabled = is_enabled = self._is_all_axes_enabled()
-            while not stop_event.is_set():
-                new_is_enabled = self._is_all_axes_enabled()
-                if new_is_enabled != is_enabled:
-                    is_enabled = new_is_enabled
-                    self.update_AxisSystemState("Enabled" if is_enabled else "Disabled")
-                time.sleep(0.1)
-
-        def update_axes_in_fault_state(stop_event: Event):
-            new_axes_in_fault = axes_in_fault = self._get_axes_in_fault_state()
-            while not stop_event.is_set():
-                new_axes_in_fault = self._get_axes_in_fault_state()
-                if new_axes_in_fault != axes_in_fault:
-                    axes_in_fault = new_axes_in_fault
-                    self.update_AxesInFaultState(axes_in_fault)
-                time.sleep(0.1)
-
-        # initial value
-        self.update_AxisSystemState("Enabled" if self._is_all_axes_enabled() else "Disabled")
-        self.update_AxesInFaultState(self._get_axes_in_fault_state())
-
-        executor.submit(update_axis_system_state, self.__stop_event)
-        executor.submit(update_axes_in_fault_state, self.__stop_event)
+        self.run_periodically(
+            PropertyUpdater(
+                self._is_all_axes_enabled,
+                not_equal,
+                lambda val: self.update_AxisSystemState("Enabled" if val else "Disabled"),
+            )
+        )
+        self.run_periodically(
+            PropertyUpdater(
+                self._get_axes_in_fault_state,
+                not_equal,
+                self.update_AxesInFaultState,
+            )
+        )
 
     def start(self) -> None:
         self.__config = ServerConfiguration(self.parent_server.server_name, ApplicationSystem().device_config.name)
@@ -71,7 +56,6 @@ class AxisSystemControlServiceImpl(AxisSystemControlServiceBase):
 
     def stop(self) -> None:
         super().stop()
-        self.__stop_event.set()
         self.__config.axis_position_counters = {name: axis.get_position_counter() for name, axis in self.__axes.items()}
         self.__config.write()
 
@@ -117,9 +101,3 @@ class AxisSystemControlServiceImpl(AxisSystemControlServiceBase):
     def ClearFaultState(self, *, metadata: MetadataDict) -> ClearFaultState_Responses:
         for axis in self.__axes.values():
             axis.clear_fault()
-
-    def stop(self) -> None:
-        super().stop()
-        self.__stop_event.set()
-        self.__config.axis_position_counters = {name: axis.get_position_counter() for name, axis in self.__axes.items()}
-        self.__config.write()