+++ sila_cetoni/motioncontrol/axis/sila/axis_service/server.py
@@ -41,11 +41,9 @@ class Server(IOServer):
             server_uuid=server_uuid,
         )
 
-        self.axispositioncontroller = AxisPositionControllerImpl(self, axis_system, self.child_task_executor)
-        self.axissystemcontrolservice = AxisSystemControlServiceImpl(self, axis_system, self.child_task_executor)
-        self.axissystempositioncontroller = AxisSystemPositionControllerImpl(
-            self, axis_system, device_properties, self.child_task_executor
-        )
+        self.axispositioncontroller = AxisPositionControllerImpl(self, axis_system)
+        self.axissystemcontrolservice = AxisSystemControlServiceImpl(self, axis_system)
+        self.axissystempositioncontroller = AxisSystemPositionControllerImpl(self, axis_system, device_properties)
 
         self.set_feature_implementation(AxisPositionControllerFeature, self.axispositioncontroller)
         self.set_feature_implementation(AxisSystemControlServiceFeature, self.axissystemcontrolservice)